+++ sipXmediaLib/include/mp/MpPortAudioDriver.h
@@ -366,20 +366,20 @@ public:
    * Gets volume meter reading for input of given stream.
    *
    * @param stream Id of audio stream
-   * @param Current volume calculated from samples.
+   * @param volume Current volume calculated from samples.
    */
    virtual OsStatus getInputVolumeMeterReading(MpAudioStreamId stream,
-                                               unsigned int& volume,
+                                               double& volume,
                                                MP_VOLUME_METER_TYPE type) const;
 
    /**
    * Gets volume meter reading for output of given stream.
    *
    * @param stream Id of audio stream
-   * @param Current volume calculated from samples.
+   * @param volume Current volume calculated from samples.
    */
    virtual OsStatus getOutputVolumeMeterReading(MpAudioStreamId stream,
-                                                unsigned int& volume,
+                                                double& volume,
                                                 MP_VOLUME_METER_TYPE type) const;
 
    /**
